requirements.txt
asgiref==3.9.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
whitenoise==6.9.0
WTForms==3.2.1
